+++ CHANGELOG.md
@@ -7,7 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## UNRELEASED
 
-_Nothing yet.._
+### Breaking
+
+- Updated `eslint` to `7.2.0`
+- Updated `eslint-plugin-jest` to `23.13.2`
+- Updated `eslint-plugin-react-hooks` to `4.0.4`
+- Updated `prettier` to `2.0.5`
+
+### Changed
+
+- Updated `eslint-config-airbnb` to `18.2.0`
+- Updated `eslint-plugin-jsx-a11y` to `6.3.0`
 
 ## 7.0.0 - Upgraded `eslint-plugin-backpack`